query_planner_agent/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
google-cloud-bigquery
langchain-google-vertexai
langchain-core